+++ NEWS.md
@@ -1,3 +1,7 @@
+# ggscribe (development)
+
+* Improved ticks and bracket linewidth.
+
 # ggscribe 0.1.1
 
 * Updated Readme.